LZW sa hash tabelom, finalna verzija.

+++ core/lzw.h
@@ -12,37 +12,37 @@
 #include"uthash.h"
 
 typedef struct{
-    long int *arr;
-    long int len;
+    int *arr;
+    int len;
 }array;
 typedef struct{
-    long int prefix;
-    long int new_char;
+    int prefix;
+    int new_char;
 }dict_element;
 typedef struct{
     dict_element *table;
-    long int len;
-    long int allocated;
+    int len;
+    int allocated;
 
 }dictionary;
 typedef struct {
     char *key;
-    long int value;
+    int value;
     UT_hash_handle hh;
 } hash_entry;
 
-void add_entry(char *key, long int value);
-long int find_entry(char *key);
+void add_entry(char *key, int value);
+int find_entry(char *key);
 void free_hash_table();
-char* make_key(long int value1, long int value2);
+char* make_key(int value1, int value2);
 void generate_random_filename(char* filename_buffer);
 void free_dictionary(dictionary *dict) ;
 void free_array(array *a);
-void add_to_dict(dictionary *dict, long int prefix, long int new_char);
-long int return_prefix(dictionary *dict, long int code);
-long int return_new_char(dictionary *dict, long int code);
-array* print_string(dictionary *dict, long int code);
-long int return_first(dictionary *dict, long int code);
+void add_to_dict(dictionary *dict, int prefix, int new_char);
+int return_prefix(dictionary *dict, int code);
+int return_new_char(dictionary *dict, int code);
+array* print_string(dictionary *dict, int code);
+int return_first(dictionary *dict, int code);
 char* compress_lzw(char* input_file_name);
 void decompress(char* input_file_name, char *output_file_name);
 #endif //COMPRESSO_LZW_H
